Sonar Fixes
