+++ src/main/common/models/generalApplication/applicationType.ts
@@ -38,40 +38,6 @@ export enum ApplicationTypeOption {
   CONFIRM_CCJ_DEBT_PAID = 'CONFIRM_CCJ_DEBT_PAID',
 }
 
-export const selectedApplicationType: Partial<{ [key in ApplicationTypeOption]: string; }> = {
-  [ApplicationTypeOption.ADJOURN_HEARING]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_HEARING',
-  [ApplicationTypeOption.AMEND_A_STMT_OF_CASE]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_CLAIM',
-  [ApplicationTypeOption.EXTEND_TIME]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.MORE_TIME',
-  [ApplicationTypeOption.OTHER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.NOT_ON_LIST',
-  [ApplicationTypeOption.RELIEF_FROM_SANCTIONS]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RELIEF_PENALTY',
-  [ApplicationTypeOption.SETTLE_BY_CONSENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SETTLING',
-  [ApplicationTypeOption.SET_ASIDE_JUDGEMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CANCEL_JUDGMENT',
-  [ApplicationTypeOption.STAY_THE_CLAIM]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.PAUSE',
-  [ApplicationTypeOption.STRIKE_OUT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.STRIKE_OUT',
-  [ApplicationTypeOption.SUMMARY_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SUMMARY_JUDGMENT',
-  [ApplicationTypeOption.UNLESS_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.IMPOSE_SANCTION',
-  [ApplicationTypeOption.VARY_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RECONSIDER',
-  [ApplicationTypeOption.VARY_PAYMENT_TERMS_OF_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.VARY_JUDGMENT',
-  [ApplicationTypeOption.CONFIRM_CCJ_DEBT_PAID]:'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CONFIRM_YOU_PAID',
-};
-
-export const selectedApplicationTypeDescription: Partial<{ [key in ApplicationTypeOption]: string; }> = {
-  [ApplicationTypeOption.ADJOURN_HEARING]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CHANGE_HEARING_DESCRIPTION',
-  [ApplicationTypeOption.AMEND_A_STMT_OF_CASE]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CHANGE_CLAIM_DESCRIPTION',
-  [ApplicationTypeOption.EXTEND_TIME]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_MORE_TIME_DESCRIPTION',
-  [ApplicationTypeOption.OTHER]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_DO_SOMETHING_DESCRIPTION',
-  [ApplicationTypeOption.RELIEF_FROM_SANCTIONS]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_RELIEF_PENALTY_DESCRIPTION',
-  [ApplicationTypeOption.SETTLE_BY_CONSENT]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_SETTLING_DESCRIPTION',
-  [ApplicationTypeOption.SET_ASIDE_JUDGEMENT]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_CANCEL_JUDGMENT_DESCRIPTION',
-  [ApplicationTypeOption.STAY_THE_CLAIM]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_PAUSE_DESCRIPTION',
-  [ApplicationTypeOption.STRIKE_OUT]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_STRIKE_OUT_DESCRIPTION',
-  [ApplicationTypeOption.SUMMARY_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_SUMMARY_JUDGMENT_DESCRIPTION',
-  [ApplicationTypeOption.UNLESS_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_IMPOSE_SANCTION_DESCRIPTION',
-  [ApplicationTypeOption.VARY_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_RECONSIDER_DESCRIPTION',
-  [ApplicationTypeOption.VARY_PAYMENT_TERMS_OF_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECT_TYPE.ASK_VARY_JUDGMENT_DESCRIPTION',
-  [ApplicationTypeOption.CONFIRM_CCJ_DEBT_PAID]:'PAGES.GENERAL_APPLICATION.SELECT_TYPE.CONFIRM_YOU_PAID_DESCRIPTION',
-};
-
 export const GeneralApplicationTypesDisplayFromCCD: { [key in ApplicationTypeOption]: string } = {
   [ApplicationTypeOption.ADJOURN_HEARING]: 'Adjourn a hearing',
   [ApplicationTypeOption.AMEND_A_STMT_OF_CASE]: 'Amend a statement of case',
