+++ src/main/common/models/generalApplication/applicationTypeConstants/selectedApplicationType.ts
@@ -1,0 +1,18 @@
+import {ApplicationTypeOption} from "models/generalApplication/applicationType";
+
+export const selectedApplicationType: Partial<{ [key in ApplicationTypeOption]: string; }> = {
+  [ApplicationTypeOption.ADJOURN_HEARING]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_HEARING',
+  [ApplicationTypeOption.AMEND_A_STMT_OF_CASE]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CHANGE_CLAIM',
+  [ApplicationTypeOption.EXTEND_TIME]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.MORE_TIME',
+  [ApplicationTypeOption.OTHER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.NOT_ON_LIST',
+  [ApplicationTypeOption.RELIEF_FROM_SANCTIONS]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RELIEF_PENALTY',
+  [ApplicationTypeOption.SETTLE_BY_CONSENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SETTLING',
+  [ApplicationTypeOption.SET_ASIDE_JUDGEMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CANCEL_JUDGMENT',
+  [ApplicationTypeOption.STAY_THE_CLAIM]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.PAUSE',
+  [ApplicationTypeOption.STRIKE_OUT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.STRIKE_OUT',
+  [ApplicationTypeOption.SUMMARY_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.SUMMARY_JUDGMENT',
+  [ApplicationTypeOption.UNLESS_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.IMPOSE_SANCTION',
+  [ApplicationTypeOption.VARY_ORDER]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.RECONSIDER',
+  [ApplicationTypeOption.VARY_PAYMENT_TERMS_OF_JUDGMENT]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.VARY_JUDGMENT',
+  [ApplicationTypeOption.CONFIRM_CCJ_DEBT_PAID]: 'PAGES.GENERAL_APPLICATION.SELECTED_APPLICATION_TYPE.CONFIRM_YOU_PAID',
+};
