+++ src/main/services/translation/convertToCUI/cuiTranslation.ts
@@ -13,7 +13,11 @@ import {toCUIPartialAdmission, toCUIRepaymentPlanFrequency} from './convertToCUI
 import {toCUICaseProgressionHearing} from 'services/translation/convertToCUI/convertToCaseProgressionHearing';
 import {DocumentType} from 'models/document/documentType';
 import {toCUICaseProgression} from 'services/translation/convertToCUI/convertToCUICaseProgression';
-import {toCUIGenericYesNo, toCUIYesNo} from 'services/translation/convertToCUI/convertToCUIYesNo';
+import {
+  toCUIGenericYesNo,
+  toCUIGenericYesNoFromUpperCase,
+  toCUIYesNo
+} from 'services/translation/convertToCUI/convertToCUIYesNo';
 import {ClaimantResponse} from 'models/claimantResponse';
 import {
   toCUICCJRequest,
@@ -106,7 +110,7 @@ export const translateCCDCaseDataToCUIModel = (ccdClaimObj: CCDClaim): Claim => 
   claim.claimType = ccdClaim.claimType;
   claim.respondentGaAppDetails = toCUIRespondentGADetails(ccdClaim.respondentSolGaAppDetails);
   claim.generalApplications = toCUIClaimGeneralApplications(ccdClaim.generalApplications);
-  claim.activeJudgment = toCUIGenericYesNo(ccdClaim.joIsLiveJudgmentExists);
+  claim.activeJudgment = toCUIGenericYesNoFromUpperCase(ccdClaim.joIsLiveJudgmentExists);
   return claim;
 };
